chore: start dlm service by dbus
When the browser call the downloader plugin in ll-box, the
plugin will also run in the same instance with browser.
We don't want this to happen because the enviroment may not
suitable for the plugin.

So we temprary use the dbus interface which provide by AM to
start the service, make it to run in the right enviroment.

This will be updated if linglong impletements the relavent dbus
interface.

Log: